Unsquashing My Horses (My Horses Were Squashed) (#3017)
* My Horses Unsquashed
Unsquashed My Horses

* My Horses Still Unsquashed
Though the manner by which this action was taken becomes slightly different.
The pixels of the courner now are counted at Three rather than Two.
In the times during which My Horses were Squashed the number was four.
but this would cause various issues in the worlde now that my horses are unsquashed.
my horses were previously squashed.