Archaic accent (#3171)
* New Accent: Archaic Speech

Do you wanna sound like a badass knight?

* just a last one

* tweaks and trails

* forgot the plural of chemicals

* should be good now

had to add some words because it was less of a hassle than to delete the lines over two files. of note: i added soldiers as "ones at arms". it isn't accurate, but as evidenced by the mrp rules update, we're not here for HRP, we're here for fun, and personally if as a female soldier someone insistantly referred to me as a man at arms i would be pretty pissed. sure, soldier isn't a role here, but some characters have a past in the army, i'm pretty sure this will come up. lmk if i forgot anything

* Update milbrat.ftl

YOU!!!!!

---------

Co-authored-by: Kip <32859367+kipdotnet@users.noreply.github.com>