Various Gun Changes (#3470)
* The changes

it's all yaml baby

* Also this change

Forgot I wanted to increase the number of Premium Deluxes you get if you're lucky to roll this bag

* Two more changes

removed the Enforcer double-tap-- it's a tremendous amount of damage in a shotgun shell to throw around, and even with the slow speed it would probably be too good. We'll think of a new way to make the Enforcer interesting.