Ourobros fix (#3474)
* Ouroboros works again

Updated the Ouroboros shots to work with the new trigger systems from upstream.

* kill Delete On Trigger

Killed DeleteOnTrigger. We didn't need it! We didn't need it. I guess we just didn't need it!

* Easy Fix

I neglected to delete this branch from the last fix I made for this specific gun and it ended up being a bit fortunate that I hadn't-- this thing is probably going to need a bit of babying for the forseeable future.

* I added a blacklist

I figured it out and added blacklist functionality to the code. I did a blacklist instead of a whitelist because, under the umbrella of food and produce, there's a million different tags and, at the moment four tags for a blacklist is really not that bad without having to need to make any new tags at all. I can see a whitelist being useful for a "pickier eater" but I think I'd prefer to cross that bridge when we get there my brain hurts.