Stun Projector and Disabler Carbine (#3548)
* The changes

it's all yaml baby

* Also this change

Forgot I wanted to increase the number of Premium Deluxes you get if you're lucky to roll this bag

* Two more changes

removed the Enforcer double-tap-- it's a tremendous amount of damage in a shotgun shell to throw around, and even with the slow speed it would probably be too good. We'll think of a new way to make the Enforcer interesting.

* Revert Hydra Change

Reverts the Hydra to its standard rate of fire

* Disabler Work Part 1

Changes the Stun Projector to the Magnum Disabler. Still working on the Carbine.

* Disabler Carbine and Implementation

Final additions before review

* Jokes

forgot to do this too ha ha ha